LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-respond-with-response-body-with-invalid-chunk.https-expected.txt


FAIL Response with a ReadableStream having non-Uint8Array chunks should be transferred as errored assert_equals: expected "PASS" but got "FAIL: read() should be rejected"
